Removed unused code from visit-button.
